Split Syntax
To make mlMainWindow.* less cluttered, I am seperating my code into seperate Headers and CPP files.
